# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UploadPresignRequest": {
        "properties": {
          "content_type": {
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ],
            "title": "Content Type",
            "type": "string"
          },
          "sha256": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Sha256",
            "type": "string"
          },
          "size": {
            "exclusiveMinimum": 0.0,
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "content_type",
          "size",
          "sha256"
        ],
        "title": "UploadPresignRequest",
        "type": "object"
      },
      "UploadPresignResponse": {
        "properties": {
          "method": {
            "const": "PUT",
            "default": "PUT",
            "title": "Method",
            "type": "string"
          },
          "required_content_type": {
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ],
            "title": "Required Content Type",
            "type": "string"
          },
          "required_size": {
            "title": "Required Size",
            "type": "integer"
          },
          "status": {
            "const": "pending",
            "title": "Status",
            "type": "string"
          },
          "upload_expires_at": {
            "format": "date-time",
            "title": "Upload Expires At",
            "type": "string"
          },
          "upload_id": {
            "format": "uuid",
            "title": "Upload Id",
            "type": "string"
          },
          "upload_url": {
            "title": "Upload Url",
            "type": "string"
          },
          "upload_url_expires_at": {
            "format": "date-time",
            "title": "Upload Url Expires At",
            "type": "string"
          }
        },
        "required": [
          "upload_id",
          "status",
          "upload_url",
          "required_content_type",
          "required_size",
          "upload_url_expires_at",
          "upload_expires_at"
        ],
        "title": "UploadPresignResponse",
        "type": "object"
      },
      "UploadResponse": {
        "properties": {
          "content_type": {
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ],
            "title": "Content Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "declared_size": {
            "title": "Declared Size",
            "type": "integer"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "status": {
            "enum": [
              "pending",
              "ready",
              "failed",
              "delete_pending",
              "deleted"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "verified_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Size"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "id",
          "status",
          "content_type",
          "declared_size",
          "verified_size",
          "sha256",
          "width",
          "height",
          "expires_at",
          "created_at",
          "updated_at"
        ],
        "title": "UploadResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "NutriPilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/uploads/presign": {
      "post": {
        "operationId": "presign_upload_api_v1_uploads_presign_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadPresignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadPresignResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Presign Upload",
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/v1/uploads/{upload_id}": {
      "delete": {
        "operationId": "remove_upload_api_v1_uploads__upload_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Upload Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Upload",
        "tags": [
          "uploads"
        ]
      },
      "get": {
        "operationId": "read_upload_api_v1_uploads__upload_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Upload Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Upload",
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/v1/uploads/{upload_id}/complete": {
      "post": {
        "operationId": "finish_upload_api_v1_uploads__upload_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Upload Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Finish Upload",
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/v1/uploads/{upload_id}/content": {
      "put": {
        "operationId": "put_upload_content_api_v1_uploads__upload_id__content_put",
        "parameters": [
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Upload Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "token",
            "required": true,
            "schema": {
              "maxLength": 2000,
              "minLength": 40,
              "title": "Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Content-Type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Content-Type"
            }
          },
          {
            "in": "header",
            "name": "Content-Length",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Content-Length"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Upload Content",
        "tags": [
          "uploads"
        ]
      }
    },
...
  }
}